Fixing some bugs
## Computation of `vcov`
In some instances, `vcov.glm` returns a non strictly positive definite matrix, which was preventing Truncated Multivariate Normal simulations.
-> Include argument `check = FALSE` in the simulation of TMVN.

Added an argument `complete` to the inference functions to let the user choose if "aliased" coefficients should be simulated.

## Issue with `cirls.fit`
In city `"bois.usa7306"` there is only a month with both enough PM2.5 data and mortality counts. A lot of `all` is `NA`.

Error was in attributing names to `Rmat` in `cirls`. In `glm.fit`, when n < p, the R matrix from the QR decomposition is padded with zeros to ensure it is a square matrix.
-> Have included the same padding in `cirls.fit`.